Renamed register.mjs to register.js and updated imports
